pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ptc-mcp"
version = "0.1.0"
description = "Programmatic Tool Calling MCP Server for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
ptc-mcp = "ptc_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ptc_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
